+++ CHANGELOG.md
@@ -20,6 +20,7 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 - Amqp reference updated to version 1.9.1 (stable)
 - Lzf reference updated to version 1.6.6 (stable)
 - Ssh2 reference updated to version 1.1 (alpha)
+- Stomp reference updated to version 2.0.1 (stable)
 - Zip reference updated to version 1.15.1 (stable)
 - DataBase `compatinfo.sqlite` is copied in same directory (<user>\.bartlett) for both phar and non phar versions.
 - Console `db:backup` command did not used anymore the system temporary folder to save DB backup files (save in same folder as DB)